[#321] Move provider validation methods to provider_details directory
