+++ app/provider_details/validate_providers.py
