+++ app/service/rest.py
@@ -92,7 +92,8 @@ from app.models import (
 )
 from app.notifications.process_notifications import persist_notification, send_notification_to_queue
 from app.schema_validation import validate
-from app.service import statistics, service_providers
+from app.service import statistics
+from app.provider_details import validate_providers
 from app.service.service_data_retention_schema import (
     add_service_data_retention_request,
     update_service_data_retention_request
@@ -218,7 +219,7 @@ def create_service():
         raise InvalidRequest(errors, status_code=400)
     data.pop('service_domain', None)
 
-    service_providers.validate_service_providers(data)
+    validate_providers.validate_service_providers(data)
 
     # validate json with marshmallow
     service_schema.load(data)
@@ -237,7 +238,7 @@ def create_service():
 def update_service(service_id):
     req_json = request.get_json()
 
-    service_providers.validate_service_providers(req_json)
+    validate_providers.validate_service_providers(req_json)
 
     fetched_service = dao_fetch_service_by_id(service_id)
     # Capture the status change here as Marshmallow changes this later
