+++ tests/app/provider_details/test_validate_providers.py
@@ -1,9 +1,9 @@
 import uuid
 
 from app.models import ProviderDetails, EMAIL_TYPE, SMS_TYPE
-from app.service.service_providers import is_provider_valid
+from app.provider_details.validate_providers import is_provider_valid
 
-PROVIDER_DETAILS_BY_ID_PATH = 'app.service.service_providers.get_provider_details_by_id'
+PROVIDER_DETAILS_BY_ID_PATH = 'app.provider_details.validate_providers.get_provider_details_by_id'
 
 
 def test_check_provider_exists(notify_db):
@@ -39,7 +39,7 @@ def test_check_provider_is_active_and_of_correct_type(mocker):
     mocked_provider_details.active = True
     mocked_provider_details.notification_type = EMAIL_TYPE
     mocker.patch(
-        'app.service.service_providers.get_provider_details_by_id',
+        PROVIDER_DETAILS_BY_ID_PATH,
         return_value=mocked_provider_details
     )
     assert is_provider_valid(uuid.uuid4(), EMAIL_TYPE) is True
